First batch of updates by GMV-ASP: - Added velocity-only receiver dynamics option - Align EKF propagation step to Gauss-Markov formulation - Update code priorities according to commonly available signals - Match base code usage to available rover codes - Manage online base station position update (may happen in VRS set-ups when connection to server is lost) - Added doppler processing in main RTK filter - Bugfix in station_id handling in rtkrcv - Added wide-lane only ambiguity resolution (disabled by default) - Added adaptive EKF processing for stec parameters (disabled by default) - Updated recommended config files in app/consapp/[rtkrcv|rnx2rtkp]/conf dirs - Add perdiodic restart option to rnx2rtkp (disabled by default) - Added FREQ_L1L5 compilation option. Switches L2 and L5 freqs processing priority (disabled by default) - Refined max rover outage behaviour to prevent EKF numerical instabilities - Redefined rnx2rtkp base-rover timestamp matching to replicate real-time behaviour when interpolation is disabled - Updated outlier detection and exclusion algorithm - Updated GUIs to reflect updates in configuration options (windows only) - Changed NEXOBS to 10 in compilation options to accomodate for number of observations in some streams - Several out-of-bound access bugfixes - Added santize options in makefiles (commented out by default) - Other minor bug and compilation warnings fixing

+++ app/qtapp/rtknavi_qt/mondlg.cpp
@@ -371,7 +371,7 @@ void MonitorDialog::ShowRtk(void)
 	rtksvrlock(&rtksvr); // lock
 
     rtk = rtksvr.rtk;
-    thread = (unsigned long)rtksvr.thread;
+    thread = (unsigned long long)rtksvr.thread;
     cycle = rtksvr.cycle;
     state = rtksvr.state;
     rtkstat = rtksvr.rtk.sol.stat;
